# docs/openapi.yaml
openapi: 3.0.3
info:
  title: MediaPulse API
  version: 1.0.0-beta.98
  description: API surface for MediaPulse. This document covers the published HTTP surface, while the markdown files in `docs/` remain the canonical source for operational notes and endpoint semantics.
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/shows/{showId}/credits/tmdb-candidates:
    post:
      summary: Fetch TMDb credit candidates without persisting them
      tags: [Shows]
      parameters:
        - { name: showId, in: path, required: true, schema: { type: integer, format: int64 } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/ShowTmdbCreditCandidatesResponse" } } }
  /api/shows/{showId}/credits/from-tmdb:
    post:
      summary: Persist one selected TMDb show credit
      tags: [Shows]
      parameters:
        - { name: showId, in: path, required: true, schema: { type: integer, format: int64 } }
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: "#/components/schemas/ShowTmdbCreditImportRequest" } } }
      responses:
        "200":
          description: OK
          content: { application/json: { schema: { $ref: "#/components/schemas/ShowPersonCreditDto" } } }
  /api/shows/{showId}/people/{personId}:
    delete:
      summary: Remove a show credit category and freeze manual curation
      tags: [Shows]
      parameters:
        - { name: showId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: personId, in: path, required: true, schema: { type: integer, format: int64 } }
        - { name: category, in: query, required: true, schema: { type: string, enum: [CAST, DIRECTING, WRITING] } }
      responses:
        "200": { description: Removed }
...
components:
...
  schemas:
...
    MovieCreditTypeDto:
      type: string
      enum: [CAST, CREW]
...
    MovieTmdbCreditCandidateDto:
      type: object
      properties:
        personTmdbId:
          type: string
        name:
          type: string
        profileUrl:
          type: string
          nullable: true
        creditType:
          $ref: "#/components/schemas/MovieCreditTypeDto"
        department:
          type: string
          nullable: true
        job:
          type: string
          nullable: true
        characterName:
          type: string
          nullable: true
        billingOrder:
          type: integer
          nullable: true
        roleLabel:
          type: string
      required: [personTmdbId, name, creditType, roleLabel]
...
    MovieTmdbCreditImportRequest:
      type: object
      properties:
        personTmdbId:
          type: string
        creditType:
          $ref: "#/components/schemas/MovieCreditTypeDto"
        department:
          type: string
          nullable: true
        job:
          type: string
          nullable: true
        characterName:
          type: string
          nullable: true
        billingOrder:
          type: integer
          nullable: true
      required: [personTmdbId, creditType]
...
    ShowPersonCreditDto:
      type: object
      properties:
        personId:
          type: integer
          format: int64
        tmdbId:
          type: string
        name:
          type: string
        slug:
          type: string
        profileUrl:
          type: string
          nullable: true
        creditType:
          type: string
          enum: [CAST, CREW]
        department:
          type: string
          nullable: true
        job:
          type: string
          nullable: true
        characterName:
          type: string
          nullable: true
        billingOrder:
          type: integer
          nullable: true
      required: [personId, tmdbId, name, slug, creditType]
...
    ShowTmdbCreditCandidatesResponse:
      type: object
      properties:
        showId: { type: integer, format: int64 }
        candidateCount: { type: integer }
        groups:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              title: { type: string }
              items:
                type: array
                items: { $ref: "#/components/schemas/MovieTmdbCreditCandidateDto" }
            required: [id, title, items]
      required: [showId, candidateCount, groups]
    ShowTmdbCreditImportRequest:
      allOf:
        - $ref: "#/components/schemas/MovieTmdbCreditImportRequest"
...